requirements.txt
# ── Frontend ──────────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── Core API ──────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9        # multipart/form-data file uploads
aiofiles>=23.2.1               # async file I/O
pydantic>=2.7.0
pydantic-settings>=2.2.1

# ── Deep Learning (optional – skip if using mock inference only) ───────────────
tensorflow>=2.16.0             # for .h5 model loading
# torch>=2.3.0                 # alternative: uncomment if using PyTorch models

# ── Classical ML ──────────────────────────────────────────────────────────────
scikit-learn>=1.4.2
numpy>=1.26.4
pandas>=2.2.2

# ── Audio Processing ──────────────────────────────────────────────────────────
librosa>=0.10.2
soundfile>=0.12.1              # backend for librosa WAV I/O

# ── Image / MRI Processing ────────────────────────────────────────────────────
Pillow>=10.3.0
nibabel>=5.2.1                 # NIfTI MRI file support

# ── PDF Report Generation ─────────────────────────────────────────────────────
reportlab>=4.2.0

# ── Explainability (Phase 2 – comment out for Phase 1) ────────────────────────
# shap>=0.45.0
# lime>=0.2.0.1

# ── Quantum (Phase 3 – comment out for Phase 1) ───────────────────────────────
# qiskit>=1.1.0
# qiskit-machine-learning>=0.7.2
# pennylane>=0.36.0

# ── Dev / Test ────────────────────────────────────────────────────────────────
httpx>=0.27.0                  # async test client for FastAPI
pytest>=8.2.0
pytest-asyncio>=0.23.6
